requirements.txt
yara-python
python-magic-win64
pefile
PyExifTool
pandas
signify
elasticsearch>=6.0.0,<7.0.0
